requirements.txt

absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.1   
#langchain==0.2.0
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.3
langsmith==0.0.83
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.0
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.8.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pip==23.0.1
platformdirs==4.1.0
posthog==3.3.3
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pypdf==4.0.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.3.0
sentencepiece==0.1.99
setuptools==65.5.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.35.1
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.1
typer==0.9.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.27.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.42.0
widgetsnbextension==4.0.9
wrapt==1.14.1
yarl==1.9.4
zipp==3.17.0
weaviate-client

